# services/wallet-service/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Core Wallet API",
        "termsOfService": "https://docs.wallet.example/terms",
        "contact": {
            "name": "Core Wallet Team",
            "email": "hoalh2@hdbank.com.vn"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/finance/restraints": {
            "get": {
                "description": "List an account's restraints (all statuses), newest-first, keyset-paginated. Pass next_cursor as ?before_seq= for the next page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "List account restraints",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1..200, default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Keyset cursor: return rows with seq_no \u003c before_seq",
                        "name": "before_seq",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestraintListResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            },
            "post": {
                "description": "Place a hold/lien on an account; reduces available balance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Add a restraint (hold/lien)",
                "parameters": [
                    {
                        "description": "Restraint request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRestraintRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RestraintResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/finance/restraints/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Get a restraint by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Restraint id (seq_no)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestraintViewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Restraint not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/finance/restraints/{id}/release": {
            "post": {
                "description": "Release an active restraint. Body is optional (reason required only for court/tax restraints, enforced by the SP).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Release a restraint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Restraint id (seq_no)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Release payload (optional)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.ReleaseRestraintRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestraintResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Restraint not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/finance/reverse": {
            "post": {
                "description": "Reverse a completed transfer by its original `reference`. Idempotent (re-reversal returns was_already_reversed=true).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Reverse an in-book transfer",
                "parameters": [
                    {
                        "description": "Transfer reversal request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TransferReversalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TransferReversalResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Original transfer not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/finance/topup/reverse": {
            "post": {
                "description": "Reverse a completed top-up by its original `reference`. Idempotent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Reverse a top-up",
                "parameters": [
                    {
                        "description": "Top-up reversal request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TopupReversalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TopupReversalResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Original top-up not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddRestraintRequest": {
            "type": "object",
            "required": [
                "acct_no",
                "restraint_purpose",
                "restraint_type"
            ],
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "narrative": {
                    "type": "string",
                    "maxLength": 500
                },
                "pledged_amt": {
                    "type": "string"
                },
                "reference_doc": {
                    "type": "string",
                    "maxLength": 500
                },
                "restraint_purpose": {
                    "type": "string"
                },
                "restraint_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.FieldError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "RFC7807: human-readable detail (dynamic context)",
                    "type": "string",
                    "example": "available 50000 \u003c required 100000"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "errorCode": {
                    "description": "stable business error code (canonical contract)",
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "errorMessage": {
                    "description": "stable human message for this code (i18n source, safe for end-user)",
                    "type": "string",
                    "example": "The account balance is not sufficient to cover this transaction"
                },
                "errors": {
                    "description": "field-level (Berlin Group / OBIE style)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FieldError"
                    }
                },
                "instance": {
                    "description": "RFC7807: request path",
                    "type": "string",
                    "example": "/v1/finance/transfer"
                },
                "internal_code": {
                    "description": "E#### for log/alert (§5)",
                    "type": "string",
                    "example": "E4022"
                },
                "iso20022_reason_code": {
                    "description": "ISO 20022 External Status Reason (§13.2)",
                    "type": "string",
                    "example": "AM04"
                },
                "retry": {
                    "$ref": "#/definitions/dto.RetryInfo"
                },
                "status": {
                    "description": "RFC7807: HTTP status code",
                    "type": "integer",
                    "example": 422
                },
                "timestamp": {
                    "description": "RFC 3339",
                    "type": "string",
                    "example": "2026-06-06T07:30:00Z"
                },
                "title": {
                    "description": "RFC7807: short human title",
                    "type": "string",
                    "example": "Insufficient funds"
                },
                "trace_id": {
                    "description": "= X-Request-Id",
                    "type": "string",
                    "example": "5f3b2c8e-1a2b-4c3d-9e8f-0a1b2c3d4e5f"
                },
                "transaction_status": {
                    "description": "pain.002 status (§13.3)",
                    "type": "string",
                    "example": "RJCT"
                },
                "type": {
                    "description": "RFC7807: URI to error doc",
                    "type": "string",
                    "example": "https://docs.wallet.example/errors/INSUFFICIENT_FUNDS"
                }
            }
        },
...
        "dto.ReleaseRestraintRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "dto.RestraintListResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RestraintViewResponse"
                    }
                },
                "next_cursor": {
                    "description": "pass as ?before_seq= for the next page",
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "dto.RestraintResponse": {
            "type": "object",
            "properties": {
                "available_bal_after": {
                    "type": "string"
                },
                "pledged_amt": {
                    "type": "string"
                },
                "restraint_id": {
                    "type": "integer"
                },
                "status": {
                    "description": "'A' added | 'R' released",
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.RestraintViewResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "description": "empty for group-scoped",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "narrative": {
                    "type": "string"
                },
                "pledged_amt": {
                    "type": "string"
                },
                "reference_doc": {
                    "type": "string"
                },
                "removed_at": {
                    "type": "string"
                },
                "removed_by": {
                    "type": "string"
                },
                "removed_reason": {
                    "type": "string"
                },
                "restraint_id": {
                    "type": "integer"
                },
                "restraint_purpose": {
                    "type": "string"
                },
                "restraint_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "description": "A active | R released | E expired",
                    "type": "string"
                }
            }
        },
        "dto.RetryInfo": {
            "type": "object",
            "properties": {
                "after_ms": {
                    "type": "integer"
                },
                "retryable": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.TopupReversalRequest": {
            "type": "object",
            "required": [
                "reason",
                "reference"
            ],
            "properties": {
                "initiator": {
                    "type": "string",
                    "enum": [
                        "OPS_MANUAL",
                        "FRAUD",
                        "DISPUTE",
                        "SYSTEM"
                    ]
                },
                "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "reference": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "dto.TopupReversalResponse": {
            "type": "object",
            "properties": {
                "event_uuid": {
                    "type": "string"
                },
                "new_balance": {
                    "type": "string"
                },
                "reversal_tran_key": {
                    "type": "integer"
                },
                "was_already_reversed": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.TransferReversalRequest": {
            "type": "object",
            "required": [
                "reason",
                "reference"
            ],
            "properties": {
                "initiator": {
                    "type": "string",
                    "enum": [
                        "OPS_MANUAL",
                        "FRAUD",
                        "DISPUTE",
                        "SYSTEM"
                    ]
                },
                "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "reference": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "dto.TransferReversalResponse": {
            "type": "object",
            "properties": {
                "event_uuid": {
                    "type": "string"
                },
                "new_balance_from": {
                    "type": "string"
                },
                "new_balance_to": {
                    "type": "string"
                },
                "reversal_tran_key": {
                    "type": "integer"
                },
                "was_already_reversed": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}